+++ photoeditor/src/test/java/ja/burhanrashid52/photoeditor/SaveSettingsTest.java
@@ -1,7 +1,10 @@
 package ja.burhanrashid52.photoeditor;
 
+import android.graphics.Bitmap;
+
 import org.junit.Test;
 
+import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertFalse;
 import static junit.framework.TestCase.assertTrue;
 
@@ -44,5 +47,29 @@ public class SaveSettingsTest {
 
         assertFalse(saveSettings.isClearViewsEnabled());
         assertFalse(saveSettings.isTransparencyEnabled());
+    }
+
+    @Test
+    public void testDefaultCompressAndQualitySaveSettings() {
+        SaveSettings saveSettings = new SaveSettings.Builder()
+                .build();
+
+        assertEquals(saveSettings.getCompressFormat(), Bitmap.CompressFormat.PNG);
+        assertEquals(saveSettings.getCompressQuality(), 100);
+    }
+
+    @Test
+    public void testCompressAndQualitySaveSettingsValues() {
+
+        Bitmap.CompressFormat compressFormat = Bitmap.CompressFormat.PNG;
+        int compressQuality = 50;
+
+        SaveSettings saveSettings = new SaveSettings.Builder()
+                .setCompressFormat(compressFormat)
+                .setCompressQuality(compressQuality)
+                .build();
+
+        assertEquals(saveSettings.getCompressFormat(), compressFormat);
+        assertEquals(saveSettings.getCompressQuality(), compressQuality);
     }
 }