+++ CHANGELOG.md
@@ -30,3 +30,24 @@ mPhotoEditor.setFilterEffect(customEffect);
 
 ## 0.3.3
 - Fixed : Brush bug using PorterDuff.Mode.SRC_OVER #80 and PR #83
+
+## 0.4.0
+- New : Added compress quality and format in save settings
+``1
+SaveSettings saveSettings = new SaveSettings.Builder()
+      .setCompressFormat(compressFormat)
+      .setCompressQuality(compressQuality)
+      .build();
+```
+- New : Added Text style builder for add and edit text
+```
+new TextStyleBuilder()
+      .withTextColor(123)
+      .withTextSize(12f)
+      .withGravity(3)
+      .withTextFont(Typeface.DEFAULT)
+      .withBackgroundColor(321)
+      .withTextAppearance(144)
+      .applyStyle(textView);
+```
+-Removed : All Deprecated methods